python_backend/accounts/throttles.py
from rest_framework.throttling import AnonRateThrottle, UserRateThrottle

class RegisterStartThrottle(AnonRateThrottle):
    scope = "register_start"

class MFAChallengeThrottle(UserRateThrottle):
    scope = "mfa_challenge"